backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2